bookmarks: Add `jj tug` inspired `advance` command
A common recommendation for new users is for them to add a `jj tug`
alias. This change makes that alias unnecessary by adding a native
version.

The command has customization points `revsets.advance-default-to`
and `revset-aliases.closest_bookmarks(to)` to adapt to different workflows.

The default target is `@` for parity with other bookmark commands.

Possible useful revsets to use can be found in
https://github.com/jj-vcs/jj/discussions/5568#discussioncomment-12674748.

Co-authored-by: Scott Taylor <scott11x8@gmail.com>